Configure consumer tag names/prefixes
Allows us to configure the consumer names (tags) that are visible in
RabbitMQ.

Tagging with specific names is useful for us because we have over 100
queues and lots of applications communicating. If we aim to be easily
identifiable, we can do these things:

- define a routing key convention that refers to the application
  publishing the message
- define a queue name convention that identifies the consumers
- identify consumers by tag names

Out of the above three we cannot do the last one. So this changeset
enables that configuration :)

Note: this changes the default tag name from `bunny` to `hutch`.