This commit restores URI parsing functionality
In https://github.com/gocardless/hutch/pull/159 functionality was added to hutch that meant that it would respect that when using `amqps://` as the URI scheme it would use TLS and also that when not specifying a port in the URI scheme it would default to the standard protocol ports (5671 for TLS connections, 5672 for non TLS connections).

This functionality appears to have been lost since and is restored by this commit it also adds a new test case to prevent another regression.