Better log levels in the message publisher
Having the entire payload logged at the "INFO" level is not ideal due to
the following reasons:
- PII data might be contained there and there is no way to fitler it out
- Payload might be big enough that logging aggregators might struggle with
the sheer amount of data
- Since the log level is shared by reusing the rails/sinatra logger, INFO
level might be too verbose for hutch but WARN level not verbose enough for
the other.

While initially I considered implementing a feature to allow for filtering
the payload for PII or triming the output, a quick solution to all problems
listed above is to change the logging statement to DEBUG, which should only
be needed in development/testing environments.